requirements.txt
matplotlib>=3.7.0
numpy>=1.24.0
PyYAML>=6.0
